edit price field

+++ products/models.py
@@ -10,7 +10,7 @@ class Product(models.Model):
     ]
 
     name = models.CharField(max_length=200)
-    price = models.DecimalField(max_digits=10, decimal_places=2, validators=[MinValueValidator(0)])
+    price = models.BigIntegerField(validators=[MinValueValidator(0)])
     image = models.ImageField(upload_to='product_images/')  # Store image path
     category = models.ForeignKey(Category, on_delete=models.CASCADE, related_name='products')
     status = models.CharField(max_length=10, choices=STATUS_CHOICES, default='active')
